+++ GUI/Main/Main.cs
@@ -675,6 +675,16 @@ namespace CKAN
             }
         }
 
+        private void userGuideToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            Utilities.ProcessStartURL("https://github.com/KSP-CKAN/CKAN/wiki/User-guide");
+        }
+
+        private void discordToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            Utilities.ProcessStartURL("https://discord.gg/Mb4nXQD");
+        }
+
         private void reportClientIssueToolStripMenuItem_Click(object sender, EventArgs e)
         {
             Utilities.ProcessStartURL("https://github.com/KSP-CKAN/CKAN/issues/new/choose");